+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   + Added generation of custom std::error_category for C++ error codes.
   + Swift enumerations now conform to CaseIterable protocol.
   + Added "InternalSetter" FDEPL property for attributes.
-  + Added "Nullable" property support for Java.
+  + Added "Nullable" property support.
 - Bug fixes:
   + Fixed potential memory leak while returning Swift proxies.
   + Fixed the default value of String type Struct fields in Java to be an empty string.